fix: move types to nx lib
