+++ libs/vre/shared/app-string-literal/src/index.ts
@@ -3,3 +3,5 @@ export * from './lib/multi-language-textarea.component';
 export * from './lib/multi-language-input.component';
 export * from './lib/human-readable-error.pipe';
 export * from './lib/multi-languages.type';
+export * from './lib/multi-language-form-array.type';
+export * from './lib/dash-language.type';
