+++ libs/vre/shared/app-string-literal/src/lib/dash-language.type.ts
@@ -1,0 +1,5 @@
+export type DaschLanguage = 'de' | 'fr' | 'it' | 'en' | 'rm';
+
+export function isDaschLanguage(value: string): value is DaschLanguage {
+  return ['de', 'fr', 'it', 'en', 'rm'].includes(value);
+}
