+++ libs/vre/shared/app-string-literal/src/lib/multi-language-form-array.type.ts
@@ -1,0 +1,4 @@
+import { FormArray } from '@angular/forms';
+import { MultiLanguageFormControl } from './multi-language-form.service';
+
+export type MultiLanguageFormArray = FormArray<MultiLanguageFormControl>;
